Split this monstrosity up a bit
