+++ app/javascript/EventList/index.jsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { BrowserRouter } from 'react-router-dom';
-import { humanize } from 'inflected';
 import moment from 'moment-timezone';
-import arrayToSentence from 'array-to-sentence';
 import Pagination from 'react-js-pagination';
 import classNames from 'classnames';
 
 import CombinedReactTableConsumer from '../Tables/CombinedReactTableConsumer';
 import { decodeStringArray, encodeStringArray } from '../Tables/FilterUtils';
+import EventCard from './EventCard';
 import EventListCategoryDropdown from './EventListCategoryDropdown';
 import { EventListCommonDataQuery, EventListEventsQuery } from './queries.gql';
 import EventListSortDropdown from './EventListSortDropdown';
+import getSortedRuns from './getSortedRuns';
 import QueryWithStateDisplay from '../QueryWithStateDisplay';
 import { ReactRouterReactTableProvider, ReactRouterReactTableConsumer } from '../Tables/ReactRouterReactTableContext';
 import {
@@ -19,12 +19,6 @@ import {
   reactTableSortToTableResultsSort,
 } from '../Tables/TableUtils';
 import { timespanFromConvention, getConventionDayTimespans } from '../TimespanUtils';
-
-function getSortedRuns(event) {
-  return [...event.runs].sort((a, b) => (
-    moment(a.starts_at).valueOf() - moment(b.starts_at).valueOf()
-  ));
-}
 
 function decodeFilterValue(field, value) {
   if (field === 'category') {
@@ -45,93 +39,6 @@ function encodeFilterValue(field, value) {
 class EventList extends React.PureComponent {
   static propTypes = {
     basename: PropTypes.string.isRequired,
-  }
-
-  renderFirstRunTime = (event, timezoneName) => {
-    if (event.runs.length > 0) {
-      const firstRunsSorted = getSortedRuns(event).slice(0, 4);
-
-      return arrayToSentence([
-        ...firstRunsSorted.map(run => (
-          moment.tz(run.starts_at, timezoneName).format('dddd h:mma')
-        )),
-        ...(
-          event.runs.length > firstRunsSorted.length
-            ? [`${event.runs.length - firstRunsSorted.length} more`]
-            : []
-        ),
-      ]);
-    }
-
-    return 'Unscheduled';
-  }
-
-  renderEventCard = (event, sorted, timezoneName) => {
-    const formResponse = JSON.parse(event.form_response_attrs_json);
-    const authorOrOrg = formResponse.author || formResponse.organization;
-
-    return (
-      <div className="card my-4" key={event.id}>
-        <div className="card-header">
-          <h4 className="m-0">
-            <a href={`/events/${event.id}`}>{event.title}</a>
-            {
-              event.category !== 'filler'
-                ? (
-                  <small className="text-muted">
-                    {' '}
-                    &mdash;
-                    {' '}
-                    {humanize(event.category)}
-                  </small>
-                )
-                : null
-            }
-          </h4>
-          {
-            authorOrOrg
-              ? (
-                <p className="m-0">
-                  <em>
-                    by
-                    {' '}
-                    {authorOrOrg}
-                  </em>
-                </p>
-              )
-              : null
-          }
-          <p className="m-0">
-            {
-              sorted.some(sort => sort.id === 'created_at')
-                ? (
-                  <strong>
-                    Added
-                    {' '}
-                    {moment.tz(event.created_at, timezoneName).format('dddd, MMMM D, YYYY [at] h:mma')}
-                  </strong>
-                )
-                : null
-            }
-            {
-              sorted.some(sort => sort.id === 'first_scheduled_run_start')
-                ? (
-                  <strong>
-                    {this.renderFirstRunTime(event, timezoneName)}
-                  </strong>
-                )
-                : null
-            }
-          </p>
-        </div>
-
-        <div
-          className="card-body"
-          // eslint-disable-next-line react/no-danger
-          dangerouslySetInnerHTML={{ __html: event.short_blurb_html }}
-        />
-      </div>
-    );
   }
 
   renderPagination = (eventsPaginated, onPageChange, extraClasses) => {
@@ -187,11 +94,11 @@ class EventList extends React.PureComponent {
       return (
         <React.Fragment key={event.id}>
           {preamble}
-          {this.renderEventCard(
-            event,
-            sorted,
-            convention.timezone_name,
-          )}
+          <EventCard
+            event={event}
+            sorted={sorted}
+            timezoneName={convention.timezone_name}
+          />
         </React.Fragment>
       );
     });
